Move schedule.ts under src/main
