+++ src/server.ts
@@ -17,7 +17,7 @@ import { ConnectionOptions } from 'tls'
 import { defaultConfig } from './config'
 import { DB } from './db'
 import { EventsProcessor } from './ingestion/process-event'
-import { startSchedule } from './services/schedule'
+import { startSchedule } from './main/services/schedule'
 import { status } from './status'
 import { PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from './types'
 import { createPostgresPool, createRedis, delay, UUIDT } from './utils'
