+++ tests/postgres/worker.test.ts
@@ -4,10 +4,10 @@ import { mocked } from 'ts-jest/utils'
 
 import Client from '../../src/celery/client'
 import { ingestEvent } from '../../src/ingestion/ingest-event'
+import { loadPluginSchedule } from '../../src/main/services/schedule'
 import { runPlugins, runPluginsOnBatch, runPluginTask } from '../../src/plugins/run'
 import { loadSchedule, setupPlugins } from '../../src/plugins/setup'
 import { ServerInstance, startPluginsServer } from '../../src/server'
-import { loadPluginSchedule } from '../../src/services/schedule'
 import { LogLevel } from '../../src/types'
 import { delay, UUIDT } from '../../src/utils'
 import { makePiscina } from '../../src/worker/piscina'
